#1124 (#8363) - use JTS WKTReader instead of PostGIS WKTReader
