+++ deegree-core/deegree-core-geometry/src/main/java/org/deegree/geometry/io/WKTReader.java
@@ -38,20 +38,16 @@ package org.deegree.geometry.io;
 import org.apache.commons.io.IOUtils;
 import org.deegree.cs.coordinatesystems.ICRS;
 import org.deegree.geometry.Geometry;
+import org.deegree.geometry.GeometryFactory;
 import org.locationtech.jts.io.ParseException;
-import org.postgis.GeometryBuilder;
-import org.postgis.binary.BinaryWriter;
 
 import java.io.IOException;
 import java.io.Reader;
-import java.sql.SQLException;
 
 /**
  * Reads {@link Geometry} objects encoded as Well-Known Text (WKT).
  *
  * TODO re-implement without delegating to JTS TODO add support for non-SFS geometries (e.g. non-linear curves) TODO
- * TODO TODO do not go about using PostGIS for parsing the WKT, generate WKB and then parse it back using JTS TODO TODO
- * TODO repeat after me s/TODO/TODO TODO/g
  *
  * @author <a href="mailto:schneider@lat-lon.de">Markus Schneider</a>
  * @author last edited by: $Author$
@@ -78,15 +74,8 @@ public class WKTReader {
 
     public Geometry read( String wkt )
                             throws ParseException {
-        try {
-            org.postgis.Geometry g = GeometryBuilder.geomFromString( wkt );
-            byte[] bs = new BinaryWriter().writeBinary( g );
-            return WKBReader.read( bs, crs );
-        } catch ( SQLException e ) {
-            e.printStackTrace();
-            // wrap the exception nicely as to not break 172643521 API calls
-            throw new ParseException( e );
-        }
+        org.locationtech.jts.geom.Geometry jtsGeom = new org.locationtech.jts.io.WKTReader().read( wkt );
+        return new GeometryFactory().createFromJTS( jtsGeom, crs );
     }
 
 }
